test3/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
feedparser>=6.0.11
openai>=1.30.0
python-jose[cryptography]==3.3.0
bcrypt>=4.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
httpx>=0.27.0
yfinance>=0.2.40
